Switch to using Module.prepend instead of method aliasing in rake hook
